Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

AmplNLReader.AmplNLPMeta

​

AmplNLReader.AmplNLPMeta — Type

AmplNLPMeta <: AbstractNLPModelMeta

A composite type that represents the main features of the optimization problem

optimize    obj(x)
subject to  lvar ≤    x    ≤ uvar
            lcon ≤ cons(x) ≤ ucon

where x is an nvar-dimensional vector, obj is the real-valued objective function, cons is the vector-valued constraint function, optimize is either "minimize" or "maximize".

Here, lvar, uvar, lcon and ucon are vectors. Some of their components may be infinite to indicate that the corresponding bound or general constraint is not present.

AmplNLPMeta(nvar; kwargs...)

Create an AmplNLPMeta with nvar variables. The following keyword arguments are accepted:

x0: initial guess

lvar: vector of lower bounds

uvar: vector of upper bounds

nbv: number of linear binary variables

niv: number of linear non-binary integer variables

nlvb: number of nonlinear variables in both objectives and constraints

nlvo: number of nonlinear variables in objectives (includes nlvb)

nlvc: number of nonlinear variables in constraints (includes nlvb)

nlvbi: number of integer nonlinear variables in both objectives and constraints

nlvci: number of integer nonlinear variables in constraints only

nlvoi: number of integer nonlinear variables in objectives only

nwv: number of linear network (arc) variables

ncon: number of general constraints

y0: initial Lagrange multipliers

lcon: vector of constraint lower bounds

ucon: vector of constraint upper bounds

nnzo: number of nonzeros in all objectives gradients

nnzj: number of elements needed to store the nonzeros in the sparse Jacobian

nnzh: number of elements needed to store the nonzeros in the sparse Hessian

nlin: number of linear constraints

nnln: number of nonlinear general constraints

nnnet: number of nonlinear network constraints

nlnet: number of linear network constraints

lin: indices of linear constraints

nln: indices of nonlinear constraints

nnet: indices of nonlinear network constraints

lnet: indices of linear network constraints

minimize: true if optimize == minimize

nlo: number of nonlinear objectives

islp: true if the problem is a linear program

name: problem name