Reference

Contents

Reference

Contents

Index

Index

AmplNLReader.AmplMPECModel

AmplNLReader.AmplNLPMeta

AmplNLReader.AmplMPECModel — Type

AmplMPECModel

Smooth reformulation for an Ampl model with complementarity constraints.

The nonlinear program

\[\begin{aligned} min_x \quad & f(x)\\ \mathrm{s.t.} \quad & c_L ≤ c(x) ≤ c_U,\\ & g_L ≤ g(x) ⟂ x ≥ x_L, \end{aligned}\]

is reformulated as

\[\begin{aligned} min_x \quad & f(x)\\ \mathrm{s.t.} \quad & c_L ≤ c(x) ≤ c_U,\\ & g_L ≤ g(x) \\ & x_L ≤ x \\ & Diag(g(x)) x ≤ 0 \end{aligned}\]

Return the original model if it does not have complementarity constraints.

AmplNLReader.AmplNLPMeta — Type

AmplNLPMeta <: AbstractNLPModelMeta

A composite type that represents the main features of the optimization problem

optimize    obj(x)
subject to  lvar ≤    x    ≤ uvar
            lcon ≤ cons(x) ≤ ucon

where x is an nvar-dimensional vector, obj is the real-valued objective function, cons is the vector-valued constraint function, optimize is either "minimize" or "maximize".

Here, lvar, uvar, lcon and ucon are vectors. Some of their components may be infinite to indicate that the corresponding bound or general constraint is not present.

AmplNLPMeta(nvar; kwargs...)

Create an AmplNLPMeta with nvar variables. The following keyword arguments are accepted:

x0: initial guess

lvar: vector of lower bounds

uvar: vector of upper bounds

nbv: number of linear binary variables

niv: number of linear non-binary integer variables

nlvb: number of nonlinear variables in both objectives and constraints

nlvo: number of nonlinear variables in objectives (includes nlvb)

nlvc: number of nonlinear variables in constraints (includes nlvb)

nlvbi: number of integer nonlinear variables in both objectives and constraints

nlvci: number of integer nonlinear variables in constraints only

nlvoi: number of integer nonlinear variables in objectives only

nwv: number of linear network (arc) variables

ncon: number of general constraints

y0: initial Lagrange multipliers

lcon: vector of constraint lower bounds

ucon: vector of constraint upper bounds

nnzo: number of nonzeros in all objectives gradients

nnzj: number of elements needed to store the nonzeros in the sparse Jacobian

lin_nnzj: number of elements needed to store the nonzeros in the sparse Jacobian of linear constraints

nln_nnzj: number of elements needed to store the nonzeros in the sparse Jacobian of nonlinear constraints

nnzh: number of elements needed to store the nonzeros in the sparse Hessian

nlin: number of linear constraints

nnln: number of nonlinear general constraints

nnnet: number of nonlinear network constraints

nlnet: number of linear network constraints

lin: indices of linear constraints

nln: indices of nonlinear constraints

nnet: indices of nonlinear network constraints

lnet: indices of linear network constraints

minimize: true if optimize == minimize

nlo: number of nonlinear objectives

islp: true if the problem is a linear program

n_cc: number of complementarity constraints

cvar: indices of variables appearing in complementarity constraints (0 if constraint is regular)

name: problem name

sparse_jacobian: indicates whether the Jacobian of the constraints is sparse

sparse_hessian: indicates whether the Hessian of the Lagrangian is sparse

grad_available: indicates whether the gradient of the objective is available

jac_available: indicates whether the Jacobian of the constraints is available

hess_available: indicates whether the Hessian of the Lagrangian is available

jprod_available: indicates whether the Jacobian-vector product J * v is available

jtprod_available: indicates whether the transpose Jacobian-vector product J' * v is available

hprod_available: indicates whether the Hessian-vector product of the Lagrangian H * v is available